+++ core/mixer/write_frame.cpp
@@ -72,7 +72,7 @@ struct write_frame::implementation
 		visitor.end();
 	}
 
-	boost::iterator_range<uint8_t*> image_data(size_t index)
+	boost::iterator_range<uint8_t*> image_data(int index)
 	{
 		if(index >= buffers_.size() || !buffers_[index]->data())
 			return boost::iterator_range<uint8_t*>();
@@ -82,11 +82,11 @@ struct write_frame::implementation
 	
 	void commit()
 	{
-		for(size_t n = 0; n < buffers_.size(); ++n)
+		for(int n = 0; n < buffers_.size(); ++n)
 			commit(n);
 	}
 
-	void commit(size_t plane_index)
+	void commit(int plane_index)
 	{
 		if(plane_index >= buffers_.size())
 			return;
@@ -127,12 +127,12 @@ write_frame& write_frame::operator=(write_frame&& other)
 }
 void write_frame::swap(write_frame& other){impl_.swap(other.impl_);}
 
-boost::iterator_range<uint8_t*> write_frame::image_data(size_t index){return impl_->image_data(index);}
+boost::iterator_range<uint8_t*> write_frame::image_data(int index){return impl_->image_data(index);}
 audio_buffer& write_frame::audio_data() { return impl_->audio_data_; }
 const void* write_frame::tag() const {return impl_->tag_;}
 const core::pixel_format_desc& write_frame::get_pixel_format_desc() const{return impl_->desc_;}
 const std::vector<safe_ptr<device_buffer>>& write_frame::get_textures() const{return impl_->textures_;}
-void write_frame::commit(size_t plane_index){impl_->commit(plane_index);}
+void write_frame::commit(int plane_index){impl_->commit(plane_index);}
 void write_frame::commit(){impl_->commit();}
 void write_frame::set_type(const field_mode::type& mode){impl_->mode_ = mode;}
 core::field_mode::type write_frame::get_type() const{return impl_->mode_;}
